// src/forgetrustee/frontend/src/types/generated/metadataSectionDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { MetadataHintDto } from './metadataHintDto';
import type { MetadataSectionDtoFields } from './metadataSectionDtoFields';
import type { MetadataSectionDtoSubsections } from './metadataSectionDtoSubsections';
import type { VisibilityRule } from './visibilityRule';

export interface MetadataSectionDto {
  /** Unique identifier for this section (machine-readable) */
  id: string;
  /** The human-readable label to show in the UI */
  label: string;
  heading?: string;
  tooltip?: string;
  helpText?: string;
  callToAction?: string;
  examples?: string[];
  hints?: MetadataHintDto[];
  visibleBy?: VisibilityRule[];
  subsections?: MetadataSectionDtoSubsections;
  fields?: MetadataSectionDtoFields;
}